+++ sql-to-dbsp-compiler/lib/sqllib/src/lib.rs
@@ -13,6 +13,7 @@ pub mod timestamp;
 pub use geopoint::GeoPoint;
 pub use interval::LongInterval;
 pub use interval::ShortInterval;
+use num_traits::Float;
 pub use source::{SourcePosition, SourcePositionRange};
 pub use timestamp::Date;
 pub use timestamp::Time;
@@ -905,13 +906,22 @@ pub fn tan_d(value: F64) -> F64 {
 some_polymorphic_function1!(tan, d, F64, F64);
 
 #[inline(always)]
+pub fn sec_d(value: F64) -> F64 {
+    (1.0 / value.into_inner().cos()).into()
+}
+
+some_polymorphic_function1!(sec, d, F64, F64);
+
+#[inline(always)]
+pub fn csc_d(value: F64) -> F64 {
+    (1.0 / value.into_inner().sin()).into()
+}
+
+some_polymorphic_function1!(csc, d, F64, F64);
+
+#[inline(always)]
 pub fn cot_d(value: F64) -> F64 {
-    let tan = value.into_inner().tan();
-    if tan.is_zero() {
-        f64::INFINITY.into()
-    } else {
-        (1.0_f64 / tan).into()
-    }
+    (1.0_f64 / value.into_inner().tan()).into()
 }
 
 some_polymorphic_function1!(cot, d, F64, F64);
@@ -967,6 +977,80 @@ pub fn cbrt_d(value: F64) -> F64 {
 }
 
 some_polymorphic_function1!(cbrt, d, F64, F64);
+
+/////////// Hyperbolic Fucntions //////////////
+
+#[inline(always)]
+pub fn sinh_d(value: F64) -> F64 {
+    value.into_inner().sinh().into()
+}
+
+some_polymorphic_function1!(sinh, d, F64, F64);
+
+#[inline(always)]
+pub fn cosh_d(value: F64) -> F64 {
+    value.into_inner().cosh().into()
+}
+
+some_polymorphic_function1!(cosh, d, F64, F64);
+
+#[inline(always)]
+pub fn tanh_d(value: F64) -> F64 {
+    value.into_inner().tanh().into()
+}
+
+some_polymorphic_function1!(tanh, d, F64, F64);
+
+#[inline(always)]
+pub fn coth_d(value: F64) -> F64 {
+    (1.0 / value.into_inner().tanh()).into()
+}
+
+some_polymorphic_function1!(coth, d, F64, F64);
+
+#[inline(always)]
+pub fn asinh_d(value: F64) -> F64 {
+    value.into_inner().asinh().into()
+}
+
+some_polymorphic_function1!(asinh, d, F64, F64);
+
+#[inline(always)]
+pub fn acosh_d(value: F64) -> F64 {
+    if value.into_inner() < 1.0 {
+        panic!("input ({}) out of range [1, Infinity]", value)
+    }
+
+    value.into_inner().acosh().into()
+}
+
+some_polymorphic_function1!(acosh, d, F64, F64);
+
+#[inline(always)]
+pub fn atanh_d(value: F64) -> F64 {
+    let inner = value.into_inner();
+    if !(-1.0..=1.0).contains(&inner) && !inner.is_nan() {
+        panic!("input ({}) out of range [-1, 1]", value)
+    }
+
+    inner.atanh().into()
+}
+
+some_polymorphic_function1!(atanh, d, F64, F64);
+
+#[inline(always)]
+pub fn csch_d(value: F64) -> F64 {
+    (1.0 / value.into_inner().sinh()).into()
+}
+
+some_polymorphic_function1!(csch, d, F64, F64);
+
+#[inline(always)]
+pub fn sech_d(value: F64) -> F64 {
+    (1.0 / value.into_inner().cosh()).into()
+}
+
+some_polymorphic_function1!(sech, d, F64, F64);
 
 ////////////////////////////////////////////////