+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SQL: Functions on binary string (octet_length, position, overlay, substring) ([#1264](https://github.com/feldera/feldera/pull/1264))
 -  pipeline-manager: add PUT endpoints for Programs, Pipelines, and Connectors (#1248)
 -  Documentation: Adding a markdown page for videos (#1126)
+- SQL: Add hyperbolic functions ([#1280](https://github.com/feldera/feldera/pull/1280))
 
 ### Fixed
 - pipeline-manager: fix a resource usage problem with http streaming under high load